alfred: Allow to specify batman-adv interface
Alfred uses the batman-adv interface to translate mac addresses and to query
the tq values for server selection. Therefore it has to know which batman-adv
interface it has to use when multiple interfaces exist. This is also necessary
when it wasn't started directly on that interface and instead on a virtual
interface which uses the batman-adv device as its slave.

Alfred assumes by default that the interface is called bat0. This name can be
changed using the parameter -b followed by the name of the interface. It is
also possible to disable the batman-adv support completely by using the special
interface name "none". This special mode is only for debugging and should not
be used in production setups.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>