alfred: Avoid buffer overruns with large MTUs
Different functions like the sending function assume an MTU smaller than the
default buffer size of 9000. This must not be true and therefore it can happend
that the buffer on the stack is overwritten.

The mtu is reduce instead to avoid this problem until all functions using such
a buffer are sanatized.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>