alfred: Change transport protocol to link-local UDP/IPv6
The main purpose of alfred is to transfer additional data between different
batman-adv nodes. Therefore using a transport layer instead of the data layer
seems to be the right choice. This allows to use the functionality like
fragmentation or checksumming without implementing it again.

It makes it also possible to reduce the privileges to non-root users
capabilities to mitigate attacks against this daemon.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>