alfred: Don't block server when waiting for master reply
A slave server should not block completely when a unix client requested a data
type and the master server doesn't immediately finishes his transaction.

Instead the unix client is stored as part of the transaction and either the
finishing or purging of the transaction triggers a reply. This also removes the
extra server loop in the unix socket code and allows to increase the data
retrieval timeout from 1 second to 10 seconds.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>