alfred: in recv_alfred_packet() add break at the end of the case block
In recv_alfred_packet() a "break" statement is missing at the end of a case
block and the function is returning -1 instead of 0 even if the packet type was
known. Add a break to let the function behave correctly.

However this bug is harmless at the moment because the return value of the
function is not checked anywhere.

Signed-off-by: Antonio Quartulli <antonio@open-mesh.com>
Signed-off-by: Simon Wunderlich <siwu@hrz.tu-chemnitz.de>